refactor String extension to DateWrangler
The other require methods were not working and DateWranger was reliant
on the Base String class - if there’s a better way to do this I am all
ears
