+++ lib/bothan/extensions/date_wrangler.rb
@@ -1,5 +1,3 @@
-require 'bothan/extensions/string'
-
 class DateWrangler
   def initialize left, right
     @left = left
@@ -70,4 +68,24 @@ class DateWrangler
       @failures = [failure]
     end
   end
+end
+
+class String
+  def titleise
+    self.split(' ').map { |w| "#{w[0].upcase}#{w[1..-1]}" }.join ' '
+  end
+
+  def to_seconds
+    ISO8601::Duration.new(self).to_seconds.seconds
+  end
+
+  def to_datetime
+    return nil if self == '*'
+    DateTime.parse self
+  end
+
+  def is_duration?
+    !self.match(/^P/).nil?
+  end
+
 end
